dtype fix

+++ p3-3.py
@@ -335,7 +335,9 @@ while True:
         device.contrast(0)
         with canvas(device) as draw:
             draw.rectangle(block, fill="red")
-        o_s.write(np.zeros(2,8820))
+        #_z = np.zeros(2,8820)
+        #_z = _z.astype(np.float32)
+        o_s.write(np.zeros(2,8820).astype(np.float32))
             
     device.clear()
